+++ LIASite/database/migrations/2024_04_16_140615_create_favorites_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('favorites', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id');
-            $table->integer('panel_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('panel_id');
             $table->timestamps();
             $table->foreign('user_id')->references('id')->on('users');
             $table->foreign('panel_id')->references('id')->on('panels');
